DataTransfer_TO_Spanner/requirements.txt
cachetools==5.3.0
colorama==0.4.6
google-auth==2.16.2
google-cloud==0.34.0
loguru==0.6.0
numpy==1.24.2
pandas==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.7.1
rsa==4.9
six==1.16.0
win32-setctime==1.1.0
